prueba heroku

+++ src/main/java/com/operation/api/models/SatoSatellite.java
@@ -1,9 +1,7 @@
 package com.operation.api.models;
 
-import models.Position;
-
 public class SatoSatellite extends Satellite {
     public  SatoSatellite (float distance){
-        super(new Position(500, 100), distance);
+        super(new Position((float)500, (float)100), distance);
     }
 }
