+++ src/main/java/com/operation/api/models/SkywalkerSatellite.java
@@ -1,9 +1,7 @@
 package com.operation.api.models;
 
-import models.Position;
-
 public class SkywalkerSatellite extends Satellite {
     public SkywalkerSatellite(float distance){
-        super(new Position(100,-100), distance);
+        super(new Position((float)100,(float)-100), distance);
     }
 }
